Refactor rate limiter filter: integrate `validate_range`, adjust field types, and update server setup logic

+++ snakeway-core/src/conf/types/runtime/listener/rate_limiter_filter.rs
@@ -11,8 +11,8 @@ pub struct RateLimiterFilterConfig {
 impl From<RateLimiterFilterSpec> for RateLimiterFilterConfig {
     fn from(spec: RateLimiterFilterSpec) -> Self {
         Self {
-            max_connections_per_second: spec.max_connections_per_second,
-            reaction_interval: Duration::from_secs(spec.reaction_interval_in_seconds),
+            max_connections_per_second: spec.max_connections_per_second as f64,
+            reaction_interval: Duration::from_secs(spec.reaction_interval_in_seconds as u64),
         }
     }
 }
